backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
pydantic==2.6.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Testing dependencies
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0
pytest-cov==5.0.0
factory-boy==3.3.0
freezegun==1.5.1 